services/lexical/requirements.txt
coverage==4.5.1
flake8===3.5.0
Flask==1.0.2
flask-cors==3.0.6
flask-debugtoolbar==0.10.1
Flask-Testing==0.6.2
gunicorn==19.8.1
psycopg2-binary==2.7.4
requests==2.19.1
jieba==0.39
opencc-python-reimplemented==0.1.4
